rm unused struct and rename state.go to pipilineconfig.go
